+++ packages/core/navigation-next/src/components/presentational/LayoutManager/LayoutManager.js
@@ -174,12 +174,7 @@ export default class LayoutManager extends Component<
       navigationUIController,
       productNavigation,
     } = this.props;
-    const {
-      isCollapsed,
-      isPeekHinting,
-      isPeeking,
-      isResizing,
-    } = navigationUIController.state;
+    const { isCollapsed, isResizing } = navigationUIController.state;
 
     const isVisible = transitionState !== 'exited';
     const shouldDisableInteraction =
@@ -194,8 +189,6 @@ export default class LayoutManager extends Component<
       >
         <ContentNavigation
           container={containerNavigation}
-          isPeekHinting={isPeekHinting}
-          isPeeking={isPeeking}
           isVisible={isVisible}
           key="product-nav"
           product={productNavigation}