+++ packages/core/navigation-next/src/components/presentational/ScrollableSectionInner/__tests__/flow/index.js
@@ -1,9 +1,0 @@
-// @flow
-
-import React from 'react';
-import ScrollableSectionInner from '../../index';
-
-<ScrollableSectionInner>Children</ScrollableSectionInner>;
-
-// $ExpectError - children prop is missing
-<ScrollableSectionInner />;