+++ packages/core/navigation-next/src/entrypoints/ItemsRenderer.js
@@ -2,6 +2,6 @@
 
 // All entrypoints will be replaced with start with `./dist/esm/` instead of `..` folder since they're connected to the bundle
 
-import ViewRenderer from '../renderer';
+import ItemsRenderer from '../renderer';
 
-export default ViewRenderer;
+export default ItemsRenderer;